+++ src/controllers/highlights/HighlightPhrase.cpp
@@ -18,16 +18,16 @@ bool HighlightPhrase::operator==(const HighlightPhrase &other) const
 {
     return std::tie(this->pattern_, this->showInMentions_, this->hasSound_,
                     this->hasAlert_, this->isRegex_, this->isCaseSensitive_,
-                    this->soundUrl_, this->color_) ==
+                    this->soundUrl_, this->color_, this->globalHighlight_, this->channels_) ==
            std::tie(other.pattern_, other.showInMentions_, other.hasSound_,
                     other.hasAlert_, other.isRegex_, other.isCaseSensitive_,
-                    other.soundUrl_, other.color_);
+                    other.soundUrl_, other.color_, other.globalHighlight_, other.channels_);
 }
 
 HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                  bool hasAlert, bool hasSound, bool isRegex,
                                  bool isCaseSensitive, const QString &soundUrl,
-                                 QColor color)
+                                 QColor color, bool globalHighlight, std::vector<std::string> channels)
     : pattern_(pattern)
     , showInMentions_(showInMentions)
     , hasAlert_(hasAlert)
@@ -42,6 +42,8 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
              QRegularExpression::UseUnicodePropertiesOption |
                  (isCaseSensitive_ ? QRegularExpression::NoPatternOption
                                    : QRegularExpression::CaseInsensitiveOption))
+    , globalHighlight_(globalHighlight)
+    , channels_(channels)
 {
     this->color_ = std::make_shared<QColor>(color);
 }
@@ -49,7 +51,8 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
 HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
                                  bool hasAlert, bool hasSound, bool isRegex,
                                  bool isCaseSensitive, const QString &soundUrl,
-                                 std::shared_ptr<QColor> color)
+                                 std::shared_ptr<QColor> color, bool globalHighlight,
+                                 std::vector<std::string> channels)
     : pattern_(pattern)
     , showInMentions_(showInMentions)
     , hasAlert_(hasAlert)
@@ -65,6 +68,8 @@ HighlightPhrase::HighlightPhrase(const QString &pattern, bool showInMentions,
              QRegularExpression::UseUnicodePropertiesOption |
                  (isCaseSensitive_ ? QRegularExpression::NoPatternOption
                                    : QRegularExpression::CaseInsensitiveOption))
+    , globalHighlight_(globalHighlight)
+    , channels_(channels)
 {
 }
 
@@ -121,6 +126,16 @@ const QUrl &HighlightPhrase::getSoundUrl() const
 const std::shared_ptr<QColor> HighlightPhrase::getColor() const
 {
     return this->color_;
+}
+
+bool HighlightPhrase::isGlobalHighlight() const
+{
+    return this->globalHighlight_;
+}
+
+const std::vector<std::string> &HighlightPhrase::getChannels() const
+{
+    return this->channels_;
 }
 
 }  // namespace chatterino