+++ src/widgets/helper/EditableModelView.cpp
@@ -8,13 +8,14 @@
 #include <QPushButton>
 #include <QTableView>
 #include <QVBoxLayout>
-
 #include <QLabel>
+#include <QTableView>
 
 namespace chatterino {
 
 EditableModelView::EditableModelView(QAbstractTableModel *model, bool movable)
     : tableView_(new QTableView(this))
+    , selectChannel(new QPushButton())
     , model_(model)
 {
     this->model_->setParent(this);
@@ -57,6 +58,12 @@ EditableModelView::EditableModelView(QAbstractTableModel *model, bool movable)
 
         for (auto &&row : rows)
             model_->removeRow(row);
+
+        selected = this->getTableView()->selectionModel()->selectedRows(0);
+        if (selected.size() == 0)
+        {
+            disableSelectChannelButton();
+        }
     });
 
     if (movable)
@@ -119,6 +126,28 @@ QAbstractTableModel *EditableModelView::getModel()
 void EditableModelView::addCustomButton(QWidget *widget)
 {
     this->buttons_->addWidget(widget);
+}
+
+void EditableModelView::addSelectChannelHighlight()
+{
+    this->selectChannel->setText("Select channels");
+    this->selectChannel->setEnabled(false);
+
+    this->buttons_->addWidget(this->selectChannel);
+
+    QObject::connect(selectChannel, &QPushButton::clicked, [this] {
+        this->selectChannelPressed.invoke();
+    });
+}
+
+void EditableModelView::disableSelectChannelButton()
+{
+    this->selectChannel->setEnabled(false);
+}
+
+void EditableModelView::enableSelectChannelButton()
+{
+    this->selectChannel->setEnabled(true);
 }
 
 void EditableModelView::addRegexHelpLink()