+++ src/widgets/helper/EditableModelView.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
 #include <QWidget>
+#include <QPushButton>
 
 #include <pajlada/signals/signal.hpp>
 
 class QAbstractTableModel;
 class QTableView;
 class QHBoxLayout;
+class QPushButton;
 
 namespace chatterino {
 
@@ -19,10 +21,15 @@ public:
 
     QTableView *getTableView();
     QAbstractTableModel *getModel();
+    QPushButton *selectChannel;
 
     pajlada::Signals::NoArgSignal addButtonPressed;
+    pajlada::Signals::NoArgSignal selectChannelPressed;
 
     void addCustomButton(QWidget *widget);
+    void addSelectChannelHighlight();
+    void disableSelectChannelButton();
+    void enableSelectChannelButton();
     void addRegexHelpLink();
 
 private: